DEBUG, DEFUNCT, refactor(wasm): add provider chunks, debug test error

+++ src/poly.rs
@@ -26,7 +26,7 @@ pub const MODQ: fn(u16) -> u16 = |x| {
 
 pub fn poly_z3_to_zq(r: &mut Poly) {
     for i in 0..NTRU_N {
-        r.coeffs[i] = r.coeffs[i] | (0 - (r.coeffs[i] >> 1)) & (NTRU_Q - 1) as u16;
+        r.coeffs[i] = r.coeffs[i] | (0 as i16 - (r.coeffs[i] as i16 >> 1)) as u16 & (NTRU_Q - 1) as u16;
     }
 }
